Initial import of plus-one logic
The logic consists of two steps.  The first is adding a relation from
the current user to the object which he voted up.  The second is
calculating how many votes the object has so far (for easy output). This
setup should make it easy to communicate the voting information with
mu-cl-resources.

Notes:

There's currently no way to downvote (or anything similar).  The service
could be extended with this sort of information, but that would greatly
chang ethe datamodel.

The predicates have not been decided on yet.  They will likely change
over time.

With the current version of the base template, escaping user-supplied
content versus SPARQL-injection is not trivial.  Verification should
happen, and transformations should be supplied.