Alter count query for correct results
It seemed our previous query did not yield the correct amount of
results.  It doesn't immediately make sense to me why the previous query
did not yield the right amount of results, but this one does the trick.
Deserves some thinking to discover whether it's a reasoning error on my
end, or a bug in the query engine.